lab4-fix: add some new mechanisms for server
1. Raft.Snapshot() will reject to create an outdated Snapshot.

  This will fix a strange case that the server call Raft.Snapshot() with
a less up-to-date index and another Raft server send a RPC call
Raft.InstallSnapshot() with a more up-to-date snapshot. If Raft do not
reject to create an outdated Snapshot, that will lead to many "index out
of range" error since more up-to-date snapshot delete much more entries
than the lower one.

2. Raft.ApplyQueue.Dequeue() will return the dequeued value.

  This will make apply thread interacting with applyQueue in an "atomic"
operation. Because apply thread is a lock free one, this fix is
necessary.

3. Introduce a new mechanism to stop enqueuing apply message.

  I use an atomic value to achieve this. Before Raft.InstallSnapshot()
enqueues a snapshot apply message to the applyQueue, we set this value
as 1. We will reset this value to 0 only after apply thread has applied
this snapshot to the server. When the thread which enqueues normal
entries to the queue see this value is 1, it wouldn't enqueue new entry
to the applyQueue.