lab4-refactor: new implementation for server and client
Almost rewrite the all code

Now server would not wait for a request which perhaps in committing, but
Start a new entry. The duplicate request check will be in the appling
phase.

The cost of CPU resource is much higher than expect, maybe there will be
something wrong should be fix.