lab5-feat: new implementation for movement
Change the implementation of movement.

The former approach is to let each server in a group independly ask
another group for imformation about movement.  This approach worked
great except for Challenge 1. The problem was that it was hard to
determine whether all servers in a group had finished movement because
leader knew nothing about followers.

For example, in the old approach, the following sequence could lead to
an error:

1. The leader detects config changes.
2. The leader starts ChangeConfig Operation.
3. The Operation is committed.
4. The leader applies ChangeConfig Operation, sends RequestMapAndSession
   to another group, and successfully update its config.
5. Senders who send entries to this group, try to delete unneeded
   shards.  So they ask leader, "Have you updated your config?".
6. The leader reply with "yes" because it has indeed updated its config.
   The senders then delete shards which they believe are no longer
   needed.
7. **HOWEVER**, a follower applies ChangeConfig Operation, send the
   RequestMapAndSession to some group but gets nothing because shards
   have already been deleted.

So new approach will make movement information(entries and sessions) is
added to Raft log. Only after this is committed, leader can reply "yes"
to another groups.